Implementation of jbs command
Now we have an array of a struct type that contains the main info of each background command that is running.
Every time that happens any input, the background command info is updated.
In the end of each iteration of the code, if any command is finished, is marked as trash info to be able to overwrite that space for an other command.